combine golu into fractional_activations
